Try using uint32_t

+++ flatgfa-c/example/example.c
@@ -1,5 +1,6 @@
 #include <stdint.h>
 #include <stdio.h>
+#include <stdint.h>
 #include <assert.h>
 #include "../include/flatgfa.h"
 
@@ -14,16 +15,16 @@ int main(int argc, const char** argv) {
     FlatGFARef* gfa = flatgfa_parse(filename);
 
     // Traverse all the paths.
-    uintptr_t path_count = flatgfa_path_count(gfa);
-    for (int i = 0; i < path_count; ++i) {
+    uint32_t path_count = flatgfa_path_count(gfa);
+    for (uint32_t i = 0; i < path_count; ++i) {
         // Print the path name.
         uintptr_t name_len;
         const uint8_t* name = flatgfa_get_path_name(gfa, i, &name_len);
         printf("%.*s:\n", name_len, name);
 
         // Traverse the steps in the path.
-        uintptr_t step_count = flatgfa_get_path_step_count(gfa, i);
-        for (int j = 0; j < path_count; ++j) {
+        uint32_t step_count = flatgfa_get_path_step_count(gfa, i);
+        for (uint32_t j = 0; j < path_count; ++j) {
             FlatGFAHandle step;
             assert(flatgfa_get_step(gfa, i, j, &step));
